99_chatbot/bedrock_text_generation_inference_server/requirements.txt
fastapi
uvicorn[standard]
boto3-1.26.140-py3-none-any.whl
botocore-1.29.140-py3-none-any.whl
